node1-recorder/requirements.txt
pyaudio==0.2.13
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0